+++ iosapp/library/HAMCardEditorViewController.h
@@ -50,7 +50,7 @@
 // for use by DayYue, not necessarily initialized
 @property BOOL addCardOnCreation; // supposed to be 0 by default
 @property (strong, nonatomic) NSString *parentID;
-@property int index;
+@property NSInteger index;
 
 - (IBAction)recordButtonTapped:(id)sender;
 - (IBAction)shootImageButtonPressed:(id)sender;
