+++ iosapp/settings/HAMEditCardPopoverViewController.h
@@ -15,13 +15,12 @@
 
 @interface HAMEditCardPopoverViewController : UIViewController<HAMCardEditorViewControllerDelegate>
 {
-
 }
 
 @property HAMSettingsViewController* mainSettingsViewController_;
 @property HAMConfig* config_;
 @property NSString* parentID_;
-@property int childIndex_;
+@property NSInteger childIndex_;
 @property UIPopoverController* popover;
 
 @property (weak, nonatomic) IBOutlet UIButton *cancelButton;
